+++ CHANGES.md
@@ -1,5 +1,7 @@
 # Unreleased
 
+- [Wrapper] Fix naming conflicts between model projections and function arguments
+  [\#324](https://github.com/ocaml-gospel/ortac/pull/324)
 - [Wrapper] Add header message in generated files with wrapper mode
   [\#322](https://github.com/ocaml-gospel/ortac/pull/322)
 - [Dune] Add automatic dune file generation to the wrapper plugin